rag/pyproject.toml
[tool.poetry]
name = "rag"
version = "1.0.0"
description = "RAG application for AIRP (AI Rate My Professor)."
authors = ["Muhammad Bilal Khan <lofi.audit@gmail.com>"]
readme = "README-REDIRECT.md"

[tool.poetry.dependencies]
python = "^3.12"
groq = "^0.9.0"
openai = "^1.41.1"
python-dotenv = "^1.0.1"
pinecone-client = {extras = ["grpc"], version = "^5.0.1"}
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
